Fix two things the Windows run found, both real on Linux too
The suite compiled and ran there for the first time and reported 150 failures.
Neither was a fault in the port.

squared_distance held the pixel gaps in a long int. That is 64 bits on Linux
and 32 on Windows, and a gap only has to reach about forty-six thousand before
its square passes what 32 bits hold -- where it wrapped negative, so far-apart
points read as neighbours and the k-d tree disagreed with brute force. The
world is 2^28 pixels across at the deepest zoom, so the gaps that do it are
ordinary ones. It is int64_t now, and the case has a test that fails on either
platform if the width is ever narrowed again.

The other is the calendar. timegm returns the negative second count for an
instant before 1970; Microsoft's _mkgmtime and gmtime_s refuse it and answer
-1. stats.c steps back through the calendar a year at a time and asserts that
distance is the inverse of step, which on Windows stopped holding as soon as
the walk crossed the epoch -- and a GPX with a broken clock is dated there,
which the code already says in two comments.

Rather than pick one of the two behaviours, both directions are now integer
arithmetic over the proleptic Gregorian calendar, so the platforms agree by
construction and time_util.c has no #ifdef left in it. It gives the same
answers as timegm and gmtime_r for everything the suite already covered, and
the cases that motivated it -- either side of the epoch, a leap day, a century
that is not a leap year, the weekday where C's % goes negative, and the
out-of-range fields stats.c relies on being carried -- are covered now too.

+++ src/heat.c
@@ -88,9 +88,15 @@ static void build_kdtree(GpxPoint **points, int lo, int hi, int depth) {
 
 static double squared_distance(GpxPoint p1, GpxPoint p2, float mercator_x_correction) {
     // Flat projection; good enough at the radii we search over.
-    long int dx = (p2.world_x - p1.world_x) * mercator_x_correction;
-    long int dy = (p2.world_y - p1.world_y);
-    return dx * dx + dy * dy;
+    //
+    // int64_t rather than long: long is 64 bits on Linux and 32 on Windows, and
+    // a world pixel gap only has to reach about forty-six thousand before its
+    // square passes what 32 bits hold. It wrapped negative there, which made
+    // far-apart points read as neighbours -- and the world is 2^28 pixels
+    // across at the deepest zoom, so the gaps that do it are ordinary ones.
+    int64_t dx = (int64_t)((double)(p2.world_x - p1.world_x) * mercator_x_correction);
+    int64_t dy = (int64_t)(p2.world_y - p1.world_y);
+    return (double)(dx * dx + dy * dy);
 }
 
 // Mercator stretches the x axis by 1/cos(latitude), so a world pixel east-west
